feat: add new features for movement inventory

+++ libs/product/src/useCase/deleteProduct/delete-product.service.ts
@@ -1,5 +1,13 @@
 /* eslint-disable no-param-reassign */
-import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import { IInventoryRepository } from '@inventory/inventory/implementations/inventory.interface';
+import { InventoryRepository } from '@inventory/inventory/infra/typeORM/repositories/inventory.repository';
+import {
+  BadRequestException,
+  ConflictException,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { IProductRepository } from '@product/product/implementations/product.interface';
 import { ProductRepository } from '@product/product/infra/typeORM/repositories/product.repository';
 
@@ -8,6 +16,8 @@ export class DeleteProductService {
   constructor(
     @Inject(ProductRepository)
     private productRepository: IProductRepository,
+    @Inject(InventoryRepository)
+    private inventoryRepository: IInventoryRepository,
   ) {}
 
   async execute(uuid: string): Promise<void> {
@@ -18,6 +28,16 @@ export class DeleteProductService {
     }
 
     const ProductId = ProductExists.id;
+
+    const InventoryExists = await this.inventoryRepository.findMovementExists(
+      ProductId,
+    );
+
+    if (InventoryExists) {
+      throw new BadRequestException(
+        'Já existem registros de movimentação de estoque para esse produto. Ao invés de excluir você deverá desativar.',
+      );
+    }
 
     return this.productRepository.delete(ProductId);
   }
